Split out normalizeURI
